pyproject.toml
[project]
name = "ha-birthday-calendar"
version = "0.1.0"
description = "A component for Home Assistant that creates a calendar entity from a CardDAV addressbook, extracting birthdays from contact VCards."
readme = "README.md"
requires-python = ">=3.13.2"
license = "MIT"
dependencies = [
    "aiohttp>=3.13.3",
    "vobject>=0.9.9",
]

[dependency-groups]
dev = [
    "pytest-homeassistant-custom-component>=0.13.102",
]
